add changes

+++ DyKnow-Killer/hitman.py
@@ -1,8 +1,10 @@
-# TODO: Force file to run as admin like in SafeGuard
+# TODO: Add sleep statments so the user can read output
+# TODO: Test on a diffrent real app like VSCode or something (Dont delete it tho)
 # TODO: Test on mock folder and create a .exe to put in there
 
 import os, sys
 import signal
+import ctypes
 import socket
 
 clear = lambda: os.system('cls')
@@ -87,6 +89,12 @@ class sd:
 
 class driver:
 
+  def is_admin():
+    try:
+      return ctypes.windll.shell32.IsUserAnAdmin()
+    except:
+      return False
+
   def addProtected():
     file = 'protect.txt'
     if not os.path.exists(file):
@@ -112,48 +120,53 @@ class driver:
 
 # Initialization code
 if __name__ == '__main__':
-  clear()
-  print("      ----- Windows DyKnow Executor ----- \
-    \nThis program will delete crucial DyKnow files to \
-    \nrender DyKnow unable to run properly. Once ran \
-    \nyou will be unable to reinstall DyKnow unless you \
-    \npull some crafty shit like recovering the files. \n")
-  input("Press 'Enter' to start \n")
-  clear()
+  if driver.is_admin():
+    clear()
+    print("      ----- Windows DyKnow Executor ----- \
+      \nThis program will delete crucial DyKnow files to \
+      \nrender DyKnow unable to run properly. Once ran \
+      \nyou will be unable to reinstall DyKnow unless you \
+      \npull some crafty shit like recovering the files. \n")
+    input("Press 'Enter' to start \n")
+    clear()
 
-  try:
-    driver.addProtected()
-    processes = sd.getProcesses()
-    blacklisted = []
-    blacklisted.extend(sd.getDyKnowProcesses())
-    if len(blacklisted) == 0:
-      clear()
-      print("The process cant locate DyKnow's files, This program might have already been ran if so please type 1 if not type 2.")
-      q_a = input('> ')
-      driver.errorHandler(q_a)
+    try:
+      driver.addProtected()
+      processes = sd.getProcesses()
+      blacklisted = []
+      blacklisted.extend(sd.getDyKnowProcesses())
+      if len(blacklisted) == 0:
+        clear()
+        print("The process cant locate DyKnow's files, This program might have already been ran if so please type 1 if not type 2.")
+        q_a = input('> ')
+        driver.errorHandler(q_a)
 
-    for file in blacklisted:
-      if file in processes and not protectedProcesses:
-        print(f'File {file} is running as process')
-        name = file.replace('.exe', '')
-        PID = sd.get_PID(name)
-        sd.killProcess(PID)
-        print(f'Killed running process {name}.')
-        for item in sd.findDyKnowExe(file):
-          os.remove(f'C:/Program Files/DyKnow/{item}')
-          print(f'Executor deleted file {item}.')
+      for file in blacklisted:
+        if file in processes and not protectedProcesses:
+          print(f'File {file} is running as process')
+          name = file.replace('.exe', '')
+          PID = sd.get_PID(name)
+          sd.killProcess(PID)
+          print(f'Killed running process {name}.')
+          for item in sd.findDyKnowExe(file):
+            os.remove(f'C:/Program Files/DyKnow/{item}')
+            print(f'Executor deleted file {item}.')
 
-      else:
-        for item in sd.findDyKnowExe(file):
-          os.remove(f'C:/Program Files/DyKnow/{item}')
-          print(f'Executor deleted file {item}.')
+        else:
+          for item in sd.findDyKnowExe(file):
+            os.remove(f'C:/Program Files/DyKnow/{item}')
+            print(f'Executor deleted file {item}.')
 
-    print(f'\nDetected files have been removed from {socket.gethostname()}.')
-    input("Press 'Enter' to quit.")
-    sys.exit(1)
+      print(f'\nDetected files have been removed from {socket.gethostname()}.')
+      input("Press 'Enter' to quit.")
+      sys.exit(1)
 
-  except Exception as e:
-    print(f'ERROR: An unknown error was encountered. \n{e}\n')
+    except Exception as e:
+      print(f'ERROR: An unknown error was encountered. \n{e}\n')
+      sys.exit(1)
+  
+  else:
+    print('Please run this program as an administrator.')
     sys.exit(1)
 
 else:
